meles now supports basic distributed conflict resolution.
When committing a change MustGet reads will be evaluated on the leader
node to make sure that no conflicts have occurred while performing the
transaction. If a conflict is found then an error will be returned to
the node issuing the commit.

Right now this is done by a very poor time synchronization mechanic
which allows for conflicts to not resolve properly if the clocks of each
node are off.

I've also added new testing helper functions to make writing meaningful
tests much easier.